requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
monai>=1.3.0

# NLP and LangChain
langchain>=0.1.0
openai>=1.0.0
tiktoken

# Data processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
pydicom>=2.4.0

# API and web
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
streamlit>=1.28.0
requests>=2.31.0

# Evaluation and metrics
scikit-learn>=1.3.0
nltk>=3.8.0
rouge-score>=0.1.2

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0